+++ include/mutate.h
@@ -1,4 +1,4 @@
 matrix copy(matrix *m);
 void clear(matrix m);
 void replace(matrix m, double old, double rep);
-void resize(matrix m, int rows, int cols);
+void resize(matrix m, size_t rows, size_t cols);
